* Add cypress * Add simple localization test
