+++ cypress/e2e/.gitignore
@@ -1,0 +1,2 @@
+1-getting-started/
+2-advanced-examples/
