+++ package.json
@@ -4,9 +4,9 @@
     "description": "",
     "author": "ikruchkov0@gmail.com",
     "scripts": {
-        "dev": "vite",
+        "dev": "vite --port 3000",
         "build": "vite build",
-        "serve": "vite preview",
+        "serve": "vite preview --port 5000",
         "test": "vitest run",
         "coverage:run": "vitest run --coverage",
         "coverage:badge": "coverage-badge-creator",
@@ -14,7 +14,8 @@
         "prepare": "husky install",
         "format": "prettier --write .",
         "format:check": "prettier --check .",
-        "lint": "eslint src/* "
+        "lint": "eslint src/* ",
+        "cypress:open": "cypress open"
     },
     "license": "MIT",
     "devDependencies": {
@@ -31,6 +32,7 @@
         "@unocss/vite": "^0.44.7",
         "c8": "^7.12.0",
         "coverage-badge-creator": "^1.0.13",
+        "cypress": "^10.6.0",
         "eslint": "^8.22.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-solid": "^0.7.1",
