+++ src/Comp/css-classes.ts
@@ -1,0 +1,4 @@
+export const cssClasses = {
+    btnSwitchLang: 'tst-switch-lang',
+    btnFlip: 'tst-flip',
+};
