+++ src/locale.ts
@@ -1,0 +1,14 @@
+export const locale = {
+    ru: {
+        Comp: {
+            SwitchLang: 'Переключить язык',
+            Flip: 'Перевернуть',
+        },
+    },
+    en: {
+        Comp: {
+            SwitchLang: 'Switch lang',
+            Flip: 'Flip',
+        },
+    },
+};
